add a fgcolor property to freetype.Font (towards Issue #75)
To make the freetype.Font properties complete, fgcolor is added. The
fgcolor argument of methods render() and render_to() becomes optional.
A default background color property is absent because it is incompatible
with the render_to() bgcolor argument.

This commit finalizes the freetype.Font Python api for Pygame 1.9.2. Consider
it stable.