BUG: some_Rect.unionall_ip( [] ) returns a new Rect instead of None
Method Rect.unionall_ip() is documented as returning NONE. As an inplace
operation, there is no need to return any other value, even if it is ignored.
It is also a performance hit on a faster alternative to Rect.unionall().

Also correct the handling of an unlikelyl error raised by PySequence_Length()
when called in unionall_ip() and unionall().